+++ warehouse/types.go
@@ -14,31 +14,6 @@ import (
 // 	Location string
 // }
 
-type PayloadT struct {
-	BatchID                      string
-	UploadID                     int64
-	StagingFileID                int64
-	StagingFileLocation          string
-	UploadSchema                 map[string]map[string]string
-	SourceID                     string
-	SourceName                   string
-	DestinationID                string
-	DestinationName              string
-	DestinationType              string
-	DestinationNamespace         string
-	DestinationRevisionID        string
-	StagingDestinationRevisionID string
-	DestinationConfig            interface{}
-	StagingDestinationConfig     interface{}
-	UseRudderStorage             bool
-	StagingUseRudderStorage      bool
-	UniqueLoadGenID              string
-	RudderStoragePrefix          string
-	Output                       []loadFileUploadOutputT
-	LoadFilePrefix               string // prefix for the load file name
-	LoadFileType                 string
-}
-
 type ProcessStagingFilesJobT struct {
 	Upload    UploadT
 	List      []*StagingFileT
